--- .eslintignore
@@ -1,2 +1,0 @@
-/modules/books.js
-/modules/bookActions.js
