+++ index.js
@@ -1,4 +1,5 @@
-import { addContent } from './modules/bookActions.js';
+// import { addContent } from './modules/bookActions.js';
+import addContent from './modules/addContent.js';
 import navEvents from './modules/navEvents.js';
 import dateTimeFunc from './modules/time.js';
 import displayBookFromStore from './modules/displayBookStore.js';
